crisPro/requirements.txt
Django==4.2.6
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-ckeditor-5==0.2.10
django-compressor==4.4
django-js-asset==2.1.0
django-libsass==0.9
django-sass==1.1.0
libsass==0.22.0
Pillow==10.1.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
rcssmin==1.1.1
rjsmin==1.2.1
soupsieve==2.5
tzdata==2023.3
Unipath==1.1
django-cors-headers==4.3.0
django-model-utils==4.3.1
django-timestampedmodel==0.1.0
djangorestframework==3.14.0
psycopg==3.1.12
pytz==2023.3.post1
setuptools==68.2.2
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
Unipath==1.1